Maj BDD et application web

+++ Cyril/telethon_project/modification_superviseurs.php
@@ -4,33 +4,18 @@
 <script type = "text/javascript" src="newjavascript.js"></script>
 
 
-           <!-- tester si l'utilisateur est connecté -->
- <?php
-                session_start();
-                if($_SESSION['username'] !== ""){
-                    $user = $_SESSION['username'];
-                    // afficher un message
-                 ?>  <div class="alert alert-success text-center">
-                     
-                        <?php   echo "Bonjour $user, vous etes bien connecté"; ?>
-                            </div>
-             
-<?php
-                }
-            ?>
-   
-<body>
- 
-
- <?php include('container.php');?> <?php require "modification_superviseur2.php" ?>   
-    <div class ="container">
-        <h1 class =" text-black p-2 my-2 position-center"> <?=$Titre?></h1>
+<body style="background-color: #EBEBEB;">
+ <?php include('container.php');?> <?php require "modification_superviseur2.php" ?> 
+    <div class ="container" >
+        <h1 class =" text-black p-2 my-2 text-center"> <?=$Titre?></h1>
         <?=$content ?>
       </div>     
  </body>
+     
+
 <table class ="table">
         <thead>
-            <tr>
+            <tr> 
                 <th scope="col">#</th>
                 <th scope="col">Nom utilisateur</th>
                 <th scope="col">Mot de passe</th>
@@ -39,24 +24,46 @@
                 <th> <a href="Ajout_superviseur2.php" class ="btn btn-primary">Ajouter compte superviseur</a></th>
             </tr>              
         </thead>
+
+ 
+       
         <tbody>        
             <?php foreach($types as $type) : ?>
-            <?php if(!isset($_GET['modifier']) || $_GET['modifier'] !== $type['id']){?>
-            <tr>
+       <?php if(!isset($_GET['modifier']) || $_GET['modifier'] !== $type['id']){?>
+            
+         <?php if($type['roles_id'] === '1'){  ?><style>.ligne{color: green;}</style>
+         
+            <tr class ="ligne">
                 <td><?= $type['id']?></td>
                 <td><?= $type['login']?></td>
                 <td><?= $type['password']?></td>
                 <td> <a href="?modifier=<?=$type['id']?>" class="btn btn-primary" >Modifier</a></td>
                 <td> <a href="?supprimer=<?=$type['id']?>" class="btn btn-danger" id ="refresh">Supprimer</a></td>                
            </tr>  
+           
+             <?php } ?> 
+           
+            <?php if($type['roles_id'] === '2'){  ?><style>.lignes{color: blue;}</style>
+         
+            <tr class ="lignes">
+                <td><?= $type['id']?></td>
+                <td><?= $type['login']?></td>
+                <td><?= $type['password']?></td>
+                <td> <a href="?modifier=<?=$type['id']?>" class="btn btn-primary" >Modifier</a></td>
+                <td> <a href="?supprimer=<?=$type['id']?>" class="btn btn-danger" id ="refresh">Supprimer</a></td>                
+           </tr>  
+           
+             <?php } ?> 
+           
             <?php } else { ?>  
+           
         <form method="POST" action="">
-            <input type="hidden" name="modifier" value="<?= $type['id']?>"/>
+            <input type="hidden" name="modifier" value="<?= $type['id']?>" />
             <td><?= $type['id'] ?> </td>
             <td><input type = "text" name ="Nom" value = "<?= $type['login']?>" /> </td>
             <td><input type = "text" name ="Mdp" value = "<?= $type['password']?>" /> </td>
 
-            <td><input type = "submit" class = "btn btn-success" value = "valider" /> </td>
+            <td><input type = "submit" class = "btn btn-success" value ="valider" id ="refresh"/> </td>
             <td><a href="modification_superviseurs.php" class = "btn btn-dangers"> Quitter </a> </td>
         </form> 							
             <?php } ?>
@@ -64,3 +71,4 @@
         </tbody>
 </table>
  
+
